requirements.txt
python-dotenv
langchain
langchain-community
langchain-openai
python-dateutil